Fix(?) undo-redo of deleting selected glyphs

+++ Scripts/UI/selected_glyphs.gd
@@ -47,6 +47,7 @@ func signal_start_holding():
 	Event_Bus.started_holding_glyphs.emit(self.get_children())
 
 func remove_child_without_undo_redo(child, delete_after_removing = true):
+	assert(child in get_children())
 	remove_child(child)
 	if delete_after_removing:
 		child.free()
@@ -68,14 +69,26 @@ func remove_all(delete_after_removing = true):
 	if get_child_count() > 0:
 		var restore_all_children_function = get_restore_all_children_function()
 		var lambda_self = self
-		Undo_Redo.add_do_method(func(): lambda_self.remove_all_without_undo_redo(delete_after_removing))
-		Undo_Redo.add_do_method(func(): Event_Bus.stopped_holding_glyphs.emit())
-		Undo_Redo.add_do_property(self, "is_holding_glyphs", false)
-		Undo_Redo.add_do_property(self, "is_selecting_glyphs", false)
+		for child in get_children():
+			var child_name = child.name
+			Undo_Redo.add_do_method(func():
+				# TODO: is this future-proof?
+				# Will there eventually be other places that the glyph could have
+				# been reparented to that aren't contained in the Canvas node?
+				lambda_self.remove_node_from_parent_by_name_without_undo_redo(child_name, delete_after_removing))
+		Undo_Redo.add_undo_method(deselect_all)
 		Undo_Redo.add_undo_method(restore_all_children_function)
-		Undo_Redo.add_undo_method(func(): Event_Bus.started_holding_glyphs.emit())
-		Undo_Redo.add_undo_property(self, "is_holding_glyphs", is_holding_glyphs)
-		Undo_Redo.add_undo_property(self, "is_selecting_glyphs", is_selecting_glyphs)
+		if is_holding_glyphs:
+			Undo_Redo.add_do_property(self, "is_holding_glyphs", false)
+			Undo_Redo.add_undo_property(self, "is_holding_glyphs", true)
+			Undo_Redo.add_do_method(func(): Event_Bus.stopped_holding_glyphs.emit())
+			Undo_Redo.add_undo_method(func(): Event_Bus.started_holding_glyphs.emit(lambda_self.get_children()))
+		if is_selecting_glyphs:
+			Undo_Redo.add_do_property(self, "is_selecting_glyphs", false)
+			Undo_Redo.add_undo_property(self, "is_selecting_glyphs", true)
+			# Also restore this node's position so the selection appears in the right place on undo.
+			var lambda_position = Vector2(position)
+			Undo_Redo.add_undo_property(self, "position", lambda_position)
 
 func delete_all():
 	remove_all(true)
@@ -101,37 +114,47 @@ func get_restore_all_children_function():
 			restore_function.call()
 	return restore_all_children_function
 
-func get_restore_child_function(child):
+func get_restore_child_function(child, make_self_parent = false):
 	var lambda_self = self
 	var lambda_glyph_instance_scene = glyph_instance_scene
 	
 	var restore_dict = child.get_restore_dict()
+	restore_dict["position"] += position
 	
 	var restore_child_function = func restore_child_function():
 		var instance = lambda_glyph_instance_scene.instantiate()
 		instance.restore_from_dict(restore_dict)
-		lambda_self.add_child(instance)
+		if make_self_parent:
+			lambda_self.add_child(instance)
+		else:
+			assert(instance.real_parent != null)
+			instance.real_parent.add_child(instance)
 	
 	return restore_child_function
 
-func restore_child(child):
-	Undo_Redo.add_do_method(get_restore_child_function(child))
+func restore_child(child, make_self_parent = false):
+	Undo_Redo.add_do_method(get_restore_child_function(child, make_self_parent))
 	var lambda_self = self
 	Undo_Redo.add_undo_method(func(): lambda_self.remove_child_by_name_without_undo_redo(child.name))
 
-func remove_child_by_name_without_undo_redo(glyph_name, delete_after_removing = true):
-	var node_path = NodePath(glyph_name)
+func remove_node_from_parent_by_name_without_undo_redo(node_name, delete_after_removing = true):
+	get_parent().remove_descendant_by_name_without_undo_redo(node_name, delete_after_removing)
+
+func remove_child_by_name_without_undo_redo(node_name, delete_after_removing = true):
+	var node_path = NodePath(node_name)
 	#Undo_Redo.add_do_method(func(): lambda_self.get_node(node_path).reparent(new_parent)) # Doesn't preserve position on redo
 	var node = get_node(node_path)
+	assert(node != null)
 	
 	remove_child_without_undo_redo(node, delete_after_removing)
 
 
-func overwrite_hold(new_instance):
+func overwrite_hold(new_instance: Glyph_Instance):
 	Undo_Redo.add_do_method(deselect_all)
 	Undo_Redo.add_undo_method(deselect_all)
-	delete_all()
-	restore_child(new_instance)
+	if is_holding_glyphs:
+		delete_all()
+	restore_child(new_instance, true)
 	signal_start_holding_child(new_instance)
 
 
@@ -145,9 +168,10 @@ func reparent_glyph_instance_by_name(glyph_name, new_parent, old_position = null
 	var do_method = func do_method():
 		var do_node = lambda_self.get_node(node_path)
 		do_node.position = new_position
-		if new_parent != lambda_self:
-			do_node.real_parent = new_parent
-		do_node.reparent(new_parent, false)
+		if new_parent == lambda_self:
+			do_node.reparent(new_parent, false)
+		else:
+			do_node.permanent_reparent(new_parent)
 	Undo_Redo.add_do_method(do_method)
 	
 	var undo_method = func undo_method():
@@ -155,15 +179,16 @@ func reparent_glyph_instance_by_name(glyph_name, new_parent, old_position = null
 		if old_position != null:
 			undo_node.position = old_position
 		if old_real_parent != null:
-			undo_node.real_parent = old_real_parent
+			undo_node.set_real_parent(old_real_parent)
 		undo_node.reparent(lambda_self, false)
 	Undo_Redo.add_undo_method(undo_method)
 
 
 #region Glyph selection methods
 func deselect_instance(child):
-	if child.real_parent:
-		child.reparent(child.real_parent)
+	var real_parent = child.get_real_parent()
+	if real_parent != null:
+		child.reparent(real_parent)
 		child.set_is_selected(false)
 	else:
 		delete_child_without_undo_redo(child)
